relay/requirements.txt
fastapi
uvicorn[standard]
pyzmq
pydantic
pydantic-settings
